Combined attack and damage rolls into one menu. Critical logic now working!
