+++ test/test_hit_roller.py
@@ -1,7 +1,7 @@
-from hit_roller import HitRoller
+from attack_roller import AttackRoller
 
 def test_hit_roll_with_adv():
-    roller = HitRoller()
+    roller = AttackRoller()
     roller.advantage = False
     roller.roll_to_hit()
     roller.advantage = True
